refactor: move Docker tasks into the cli subpackage
Move the ten root-level tasks (Build, Compose, ImageLs, Prune, Pull, Push,
Rm, Run, Stop, Tag) and their helpers into io.kestra.plugin.docker.cli so
the plugin no longer mixes root-package and subpackage tasks (plugin-doc-lint
PKG-003). Each task keeps its previous type as a @Plugin alias, so existing
flows referencing io.kestra.plugin.docker.<Task> keep working. Adds cli
metadata and icon and updates examples and tests to the new package.
