+++ src/main/java/io/kestra/plugin/docker/cli/package-info.java
@@ -2,6 +2,6 @@
     description = "This subgroup of plugins contains tasks for interacting with the Docker API.",
     categories = { PluginSubGroup.PluginCategory.CORE, PluginSubGroup.PluginCategory.INFRASTRUCTURE }
 )
-package io.kestra.plugin.docker;
+package io.kestra.plugin.docker.cli;
 
 import io.kestra.core.models.annotations.PluginSubGroup;
