+++ src/main/resources/metadata/cli.yaml
@@ -1,0 +1,8 @@
+group: io.kestra.plugin.docker.cli
+name: "cli"
+title: "Docker"
+description: "Docker tasks for building images, running containers, and managing artifacts from Kestra workflows."
+body: "Requires access to a Docker daemon via host socket or TCP endpoint; the caller inherits daemon permissions, so scope access carefully. Tasks cover build, push/pull, tagging, Compose execution, container lifecycle, and pruning. Registry auth is optional and set per task; pass tags, compose files, and run options explicitly."
+videos: []
+createdBy: "Kestra Core Team"
+managedBy: "Kestra Core Team"
