+++ src/test/java/io/kestra/plugin/docker/cli/BuildTest.java
@@ -1,4 +1,4 @@
-package io.kestra.plugin.docker;
+package io.kestra.plugin.docker.cli;
 
 import java.nio.file.Files;
 import java.nio.file.Path;
