+++ src/test/java/io/kestra/plugin/docker/cli/ComposeTest.java
@@ -1,4 +1,4 @@
-package io.kestra.plugin.docker;
+package io.kestra.plugin.docker.cli;
 
 import java.util.List;
 import java.util.Map;
