+++ src/test/java/io/kestra/plugin/docker/cli/ImageLsTest.java
@@ -1,4 +1,4 @@
-package io.kestra.plugin.docker;
+package io.kestra.plugin.docker.cli;
 
 import java.util.Map;
