+++ src/test/java/io/kestra/plugin/docker/cli/PruneTest.java
@@ -1,4 +1,4 @@
-package io.kestra.plugin.docker;
+package io.kestra.plugin.docker.cli;
 
 import java.util.Map;
