+++ src/test/java/io/kestra/plugin/docker/cli/PullTest.java
@@ -1,4 +1,4 @@
-package io.kestra.plugin.docker;
+package io.kestra.plugin.docker.cli;
 
 import org.junit.jupiter.api.Test;
