+++ src/test/java/io/kestra/plugin/docker/cli/SanityChecks.java
@@ -1,4 +1,4 @@
-package io.kestra.plugin.docker;
+package io.kestra.plugin.docker.cli;
 
 import org.junit.jupiter.api.Test;
 
@@ -12,11 +12,12 @@ import static org.hamcrest.Matchers.hasSize;
 import static org.hamcrest.Matchers.is;
 
 @KestraTest(startRunner = true)
-class RunnerTest {
+class SanityChecks {
+
     @Test
-    @ExecuteFlow("flows/docker_compose.yaml")
-    void flow(Execution execution) {
-        assertThat(execution.getTaskRunList(), hasSize(2));
+    @ExecuteFlow("sanity-checks/docker_run_output_file.yml")
+    void runWithOutputFile(Execution execution) {
+        assertThat(execution.getTaskRunList(), hasSize(3));
         assertThat(execution.getState().getCurrent(), is(State.Type.SUCCESS));
     }
 }
