+++ src/test/resources/sanity-checks/docker_run_output_file.yml
@@ -8,7 +8,7 @@ inputs:
 
 tasks:
   - id: docker_run_output_file
-    type: io.kestra.plugin.docker.Run
+    type: io.kestra.plugin.docker.cli.Run
     containerImage: ubuntu:22.04
     commands:
       - "/bin/sh"
